requirements.txt
asgiref==3.3.1
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.6
django-crispy-forms==1.10.0
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
folium==0.11.0
fontawesome-free==5.15.3
idna==2.10
Jinja2==2.11.3
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.3
numpy==1.20.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.0.1
python-decouple==3.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.3
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
